+++ dev/ppr_tables_bbwp.R
@@ -40,7 +40,19 @@ require(data.table); require(readxl);library(usethis)
   # save measures as bbwp table
   use_data(er_crops, overwrite = TRUE)
   
-
+# -- prepare correction factors for financial reward per Agricultural Economic Region for Ecoregelingen ---
+  
+  # load in csv
+  er_aer_reward <- as.data.table(fread('dev/220519 ecoregeling reward weging.csv',dec=','))
+  
+  # convert UTF-8 encoded strings to latin1 if required
+  if('UTF-8' %in% Encoding(er_aer_reward$statname)) {
+    er_aer_reward$statname <- iconv(er_aer_reward$statname, from = '', to = 'latin1')
+  }
+  # save measures as bbwp table
+  use_data(er_aer_reward, overwrite = TRUE)
+  
+  
 # -- prepare LSW table