+++ tests/testthat/test-er.R
@@ -1,25 +1,27 @@
 require(testthat)
 
   # # default input for testing
-  # B_SOILTYPE_AGR = c('dekzand', 'loess', 'rivierklei')
-  # B_GWL_CLASS = c('GtIII', 'GtI', 'GtV')
-  # A_P_SG = c(0.4, 0.8, 1)
-  # B_SLOPE_DEGREE = c(1.5,4,1.5)
-  # B_LU_BRP = c(265, 1932, 266)
-  # B_LU_BBWP = c(1,4,1)
-  # M_DRAIN = c(TRUE, FALSE, TRUE)
-  # D_SA_W = c(0, 0.5, 1)
-  # D_AREA = c(100,80,2.5)
-  # measures = NULL
-  # farmscore = 100
-  # sector = c('dairy', 'arable')
-  # output = 'measures'
+  B_SOILTYPE_AGR = c('dekzand', 'loess', 'rivierklei')
+  B_GWL_CLASS = c('GtIII', 'GtI', 'GtV')
+  A_P_SG = c(0.4, 0.8, 1)
+  B_SLOPE_DEGREE = c(1.5,4,1.5)
+  B_AER_CBS = c('LG05','LG14','LG02')
+  B_LU_BRP = c(265, 1932, 266)
+  B_LU_BBWP = c(1,4,1)
+  M_DRAIN = c(TRUE, FALSE, TRUE)
+  D_SA_W = c(0, 0.5, 1)
+  D_AREA = c(100,80,2.5)
+  measures = NULL
+  farmscore = 100
+  sector = c('dairy', 'arable')
+  output = 'scores'
 
 # run example 1 without any measures taken
 test <- ecoregeling(B_SOILTYPE_AGR = c('dekzand', 'loess', 'rivierklei'),
                     B_GWL_CLASS = c('GtIII', 'GtI', 'GtV'),
                     A_P_SG = c(0.4, 0.8, 1),
                     B_SLOPE_DEGREE = c(1.5,4,1.5),
+                    B_AER_CBS = c('LG05','LG14','LG02'),
                     B_LU_BRP = c(265, 1932, 266),
                     B_LU_BBWP = c(1,4,1),
                     M_DRAIN = c(TRUE, FALSE, TRUE),
@@ -31,6 +33,7 @@ test <- ecoregeling(B_SOILTYPE_AGR = c('dekzand', 'loess', 'rivierklei'),
                     output = 'scores'
                    )
 
+
   # run tests on format and output values
   test_that("check ecoregeling", {
     expect_equal(
@@ -41,7 +44,7 @@ test <- ecoregeling(B_SOILTYPE_AGR = c('dekzand', 'loess', 'rivierklei'),
   test_that("check ecoregeling", {
     expect_equal(
       object = colnames(test$fields),
-      expected = c("field_id","s_er_soil","s_er_water","s_er_climate","s_er_biodiversity","s_er_landscape","s_er_tot"))
+      expected = c("field_id","s_er_soil","s_er_water","s_er_climate","s_er_biodiversity","s_er_landscape","s_er_tot","reward"))
   })
   
   test_that("check ecoregeling", {
@@ -54,7 +57,7 @@ test <- ecoregeling(B_SOILTYPE_AGR = c('dekzand', 'loess', 'rivierklei'),
   test_that("check ecoregeling", {
     expect_equal(
       object = as.numeric(unlist(test$farm)),
-      expected = c(6,9,6,11,6,0),
+      expected = c(6,9,6,11,6,0,460),
       tolerance = 0.01)
   })
 
@@ -64,7 +67,7 @@ dt.measures <- as.data.table(BBWPC::bbwp_measures)
 dt.measures <- dt.measures[!is.na(eco_id)]
 
 # make measurement list for 2 of the 4 fields
-measures <- rbind(data.table(id = 1, dt.measures[c(2,5,18,28,32,3,38,43,62)]),
+measures <- rbind(data.table(id = 1, dt.measures[c(2,5,18,28,32,3,38,43,58)]),
                   data.table(id = 3, dt.measures[c(7,21,30,46,5)]))
 
 
@@ -75,6 +78,7 @@ test <- ecoregeling(B_SOILTYPE_AGR = c('dekzand', 'loess', 'rivierklei'),
                     B_SLOPE_DEGREE = c(1.5,4,1.5),
                     B_LU_BRP = c(265, 1932, 266),
                     B_LU_BBWP = c(1,4,1),
+                    B_AER_CBS = c('LG05','LG14','LG02'),
                     M_DRAIN = c(TRUE, FALSE, TRUE),
                     D_SA_W = c(0, 0.5, 1),
                     D_AREA = c(100,80,2.5),
@@ -88,14 +92,14 @@ test <- ecoregeling(B_SOILTYPE_AGR = c('dekzand', 'loess', 'rivierklei'),
   test_that("check ecoregeling", {
     expect_equal(
       object = test$fields$s_er_tot,
-      expected = c(42,6,42),
+      expected = c(20,6,43),
       tolerance = 0.01)
   })
   
   test_that("check ecoregeling", {
     expect_equal(
       object = as.numeric(unlist(test$farm)),
-      expected = c(26,34,30,61,16,15),
+      expected = c(14,14,15,29,13,8,682),
       tolerance = 0.01)
   })
 
@@ -106,6 +110,7 @@ test <- ecoregeling(B_SOILTYPE_AGR = c('dekzand', 'loess', 'rivierklei'),
                       B_SLOPE_DEGREE = c(1.5,4,1.5),
                       B_LU_BRP = c(265, 1932, 266),
                       B_LU_BBWP = c(1,4,1),
+                      B_AER_CBS = c('LG05','LG14','LG02'),
                       M_DRAIN = c(TRUE, FALSE, TRUE),
                       D_SA_W = c(0, 0.5, 1),
                       D_AREA = c(100,80,2.5),